requirements.txt
streamlit>=1.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
jsonschema>=4.23.0
google-generativeai>=0.8.0
pytest>=8.3.0
sentence-transformers>=2.7.0  # semantic table selection (~80MB model downloaded on first run)
